+++ eliot/__init__.py
@@ -6,8 +6,8 @@ from warnings import warn
 # Expose the public API:
 from ._message import Message
 from ._action import (
-    startAction, startTask, Action, preserve_context, currentAction as
-    current_action
+    startAction, startTask, Action, preserve_context, current_action,
+    use_asyncio_context,
 )
 from ._output import (
     ILogger,
@@ -33,6 +33,7 @@ def add_destination(destination):
     Logger._destinations.add(destination)
 
 
+# Backwards compatibilty:
 addDestination = add_destination
 removeDestination = Logger._destinations.remove
 addGlobalFields = Logger._destinations.addGlobalFields
@@ -45,6 +46,7 @@ write_failure = writeFailure
 add_destinations = Logger._destinations.add
 remove_destination = removeDestination
 add_global_fields = addGlobalFields
+
 
 __all__ = [
     "Message",
@@ -67,6 +69,7 @@ __all__ = [
     "FileDestination",
     "register_exception_extractor",
     "current_action",
+    "use_asyncio_context",
 
     # PEP 8 variants:
     "write_traceback",