+++ docs/source/news.rst
@@ -1,6 +1,18 @@
 What's New
 ==========
 
+1.2.0
+^^^^^
+
+Features:
+
+* Eliot now does the right thing for ``asyncio`` coroutines in Python 3.5 and later.
+  See :ref:`asyncio_coroutine` for details.
+
+Misc:
+
+* ``Action.continue_task`` can now accept text task IDs (``str`` in Python 3, ``unicode`` in Python 2).
+
 1.1.0
 ^^^^^